Renamed util to utility, exposed type mapper
